requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
babel==2.16.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-appconf==1.0.6
django-formtools==2.5.1
django-otp==1.5.4
django-otp-sms==1.3
django-phonenumber-field==2.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.52
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
twilio==9.4.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
